Move methods from __init__ to cli
